heredoc signal and heredoc files norms are fixed

+++ executor/heredocs.c
@@ -6,104 +6,33 @@
 /*   By: aldurmaz <aldurmaz@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/05 12:58:57 by aldurmaz          #+#    #+#             */
-/*   Updated: 2025/08/20 01:53:15 by aldurmaz         ###   ########.fr       */
+/*   Updated: 2025/08/20 16:14:56 by aldurmaz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void free_for_heredoc(t_shell *head)
+static int	create_tmpfile(char *buf, size_t buf_size, int *fd)
 {
-	static t_shell *tmp;
-
-	if (head)
-	{
-		tmp = head;
-	}
-	if (head == NULL)
-	{
-		cleanup_and_exit(tmp, 130);
-	}
-}
-
-
-void setup_heredoc_signal(int signo)
-{
-	(void)signo;
-	g_status = STATUS_HEREDOC_CTRL_C;
-	write(2,"\n",1);
-	free_for_heredoc(NULL);
-	exit(130);
-}
-static void	child_heredoc_routine(t_redir *redir, int fd, t_shell *shell)
-{
-	char	*line;
-	char	*expanded;
-
-	// setup_signals(MODE_CHILD);
-	signal(SIGINT,setup_heredoc_signal);
-	while (1)
-	{
-		line = readline("> ");
-		if (!line)
-		{
-			ft_putstr_fd("minishell: warning: heredoc delimited by EOF (wanted `", 2);
-			ft_putstr_fd(redir->filename, 2);
-			ft_putstr_fd("')\n", 2);
-			break;
-		}
-		if (ft_strncmp(line, redir->filename, ft_strlen(line) + 1) == 0)
-		{
-			free(line);
-			break;
-		}
-		if (redir->expand_in_heredoc)
-		{
-			expanded = expand_heredoc_line(line, shell);
-			write(fd, expanded, ft_strlen(expanded));
-			free(expanded);
-		}
-		else
-			write(fd, line, ft_strlen(line));
-		write(fd, "\n", 1);
-		free(line);
-	}
-	cleanup_and_exit(shell, 0);
-}
-
-static int	process_single_heredoc(t_redir *redir, t_simple_command *cmd, t_shell *shell)
-{
-	pid_t	pid;
-	int		status;
-	int		fd;
 	static int	counter = 0;
-	char		tmp_filename_buf[64];
-	char		*dup_path;
 	char		*num;
-	ft_bzero(tmp_filename_buf, sizeof(tmp_filename_buf));
-	ft_strlcpy(tmp_filename_buf, "/tmp/minishell_heredoc_", sizeof(tmp_filename_buf));
+
+	ft_bzero(buf, buf_size);
+	ft_strlcpy(buf, "/tmp/minishell_heredoc_", buf_size);
 	num = ft_itoa(counter++);
 	if (!num)
 		return (-1);
-	ft_strlcat(tmp_filename_buf, num, sizeof(tmp_filename_buf));
+	ft_strlcat(buf, num, buf_size);
 	free(num);
-	fd = open(tmp_filename_buf, O_CREAT | O_WRONLY | O_TRUNC, 0600);
-	if (fd < 0)
+	*fd = open(buf, O_CREAT | O_WRONLY | O_TRUNC, 0600);
+	if (*fd < 0)
 		return (-1);
-	signal(SIGINT, SIG_IGN);
-	pid = fork();
-	if (pid == -1)
-	{
-		close(fd);
-		unlink(tmp_filename_buf);
-		return (-1);
-	}
-	if (pid == 0)
-		child_heredoc_routine(redir, fd, shell);
-	close(fd);
-	waitpid(pid, &status, 0);
-	setup_signals(MODE_INTERACTIVE); // Restore parent signal handling, burası modeparent mı olcak?
+	return (0);
+}
 
+static int	handle_heredoc_child_status(int status, char *tmp_filename_buf,
+	t_shell *shell)
+{
 	if (WIFSIGNALED(status))
 	{
 		if (WTERMSIG(status) == SIGINT)
@@ -121,24 +50,51 @@ static int	process_single_heredoc(t_redir *redir, t_simple_command *cmd, t_shell
 			return (-1);
 		}
 	}
-	else if (WIFEXITED(status))
+	else if (WIFEXITED(status) && WEXITSTATUS(status) == 130)
 	{
-		// Eğer çıkış kodu 130 ise, bu bizim Ctrl+C durumumuzdur.
-		if (WEXITSTATUS(status) == 130)
-		{
-			unlink(tmp_filename_buf); // Geçici dosyayı sil
-			shell->exit_code = 130;
-			return (-1); // Döngüyü kırmak için -1 döndür
-		}
+		unlink(tmp_filename_buf);
+		shell->exit_code = 130;
+		return (-1);
 	}
+	return (0);
+}
+
+static int	handle_heredoc_result(int status, char *tmp_filename_buf,
+		t_redir *redir, t_shell *shell)
+{
+	char	*dup_path;
+
+	setup_signals(MODE_INTERACTIVE);
+	if (handle_heredoc_child_status(status, tmp_filename_buf, shell) == -1)
+		return (-1);
 	dup_path = ft_strdup(tmp_filename_buf);
 	if (!dup_path)
 		return (-1);
 	free(redir->filename);
 	redir->filename = dup_path;
+	return (0);
+}
+
+static int	process_single_heredoc(t_redir *redir,
+	t_simple_command *cmd, t_shell *shell)
+{
+	pid_t	pid;
+	int		status;
+	int		fd;
+	char	tmp_filename_buf[64];
 
 	(void)cmd;
-	return (0);
+	if (create_tmpfile(tmp_filename_buf, sizeof(tmp_filename_buf), &fd) == -1)
+		return (-1);
+	signal(SIGINT, SIG_IGN);
+	pid = fork();
+	if (pid == -1)
+		return (close(fd), unlink(tmp_filename_buf), -1);
+	if (pid == 0)
+		child_heredoc_routine(redir, fd, shell);
+	close(fd);
+	waitpid(pid, &status, 0);
+	return (handle_heredoc_result(status, tmp_filename_buf, redir, shell));
 }
 
 int	handle_heredocs(t_command_chain *chain, t_shell *shell)
@@ -155,9 +111,10 @@ int	handle_heredocs(t_command_chain *chain, t_shell *shell)
 			r = (t_redir *)lst->content;
 			if (r && r->type == TOKEN_HEREDOC)
 			{
-				if(g_status != STATUS_HEREDOC_CTRL_C)
+				if (g_status != STATUS_HEREDOC_CTRL_C)
 				{
-					if (process_single_heredoc(r, chain->simple_command, shell) == -1)
+					if (process_single_heredoc(r, chain->simple_command,
+							shell) == -1)
 						return (-1);
 				}
 			}
@@ -166,21 +123,4 @@ int	handle_heredocs(t_command_chain *chain, t_shell *shell)
 		chain = chain->next;
 	}
 	return (0);
-}
-
-void	cleanup_heredoc_files_after_exec(t_simple_command *cmd)
-{
-	t_list	*lst;
-	t_redir	*r;
-
-	if (!cmd)
-		return ;
-	lst = cmd->redirections;
-	while (lst)
-	{
-		r = (t_redir *)lst->content;
-		if (r && r->type == TOKEN_HEREDOC && r->filename)
-			unlink(r->filename);
-		lst = lst->next;
-	}
 }
